+++ react-lesson/src/components/Article/Article.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Articles } from './articles';
+import { Articles } from './../articles/articles';
 
 
 export const Article = ({data}) => {
