--- react-lesson/src/components/articles.scss
@@ -1,3 +1,0 @@
-body {
-    // background-color: red;
-}
