+++ react-lesson/src/components/author/author.js
