+++ react-lesson/src/components/description/description.js
